Add support for building on kernels with version >= 4.20.
The kernel has deprecated support for time accessors that provide
results in structures a-la `struct timeval` or `struct timespec`.

The relevant accessors in the source tree have been removed and
replacements utilized. To avoid having to insert a significant
amount of conditional compilation macros asiring from the different
layouts of the `ktime` union on systems with different word sizes,
only 64-bit architectures will be supported.